compass, cherokee, and lightpd have been committed
due to public disinterest, lack of personal knowledge, and extraneous
use of frameworks that deter the learning of the backbone of languages,
such as CSS, SASS, and SCSS. Although SCSS and SASS are implementations
of CSS, SASS and CSS both can import Compass. Compass can be seen as a
Twitter Bootstrap, therefore has been omitted. Bootstrap will also not
be present here as it exhibits bloated and unnecessary code which is
almost 97% unused on normal sites made strictly with Bootstrap. Since
Bootstrap is a CSS framework, it will be omitted from the learning.